Bump gradle version. Fix of negotiation cordapp.
