+++ cordapp-example/build.gradle
@@ -2,7 +2,7 @@ buildscript {
     ext {
         corda_release_group = 'net.corda'
         corda_release_version = '4.0-SNAPSHOT'
-        corda_gradle_plugins_version = '4.0.36'
+        corda_gradle_plugins_version = '4.0.37'
         kotlin_version = '1.2.71'
         junit_version = '4.12'
         quasar_version = '0.7.10'
