+++ negotiation-cordapp/.idea/runConfigurations/Run_Flow_Tests.xml
@@ -1,26 +1,21 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Run Flow Tests" type="JUnit" factoryName="JUnit">
-    <module name="cordapp_test" />
+    <module name="workflows_test" />
     <extension name="coverage">
       <pattern>
         <option name="PATTERN" value="com.negotiation.flows.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" />
-    <option name="PACKAGE_NAME" value="com.negotiation.flows" />
+    <option name="PACKAGE_NAME" value="negotiation.workflows.flows" />
     <option name="MAIN_CLASS_NAME" value="" />
     <option name="METHOD_NAME" value="" />
     <option name="TEST_OBJECT" value="package" />
     <option name="VM_PARAMETERS" value="-ea -javaagent:../lib/quasar.jar" />
     <option name="PARAMETERS" value="" />
-    <option name="WORKING_DIRECTORY" value="file://%MODULE_WORKING_DIR%" />
-    <option name="PASS_PARENT_ENVS" value="true" />
-    <option name="TEST_SEARCH_SCOPE">
-      <value defaultName="singleModule" />
-    </option>
-    <patterns />
-    <method />
+    <option name="WORKING_DIRECTORY" value="%MODULE_WORKING_DIR%" />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
   </configuration>
 </component>
