+++ negotiation-cordapp/build.gradle
@@ -124,7 +124,6 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10006")
             adminAddress("localhost:10046")
         }
-        webPort 10007
         cordapps = [
                 "$project.group:contracts:$project.version",
                 "$project.group:workflows:$project.version"
@@ -138,7 +137,6 @@ task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['jar']) {
             address("localhost:10009")
             adminAddress("localhost:10049")
         }
-        webPort 10010
         cordapps = [
                 "$project.group:contracts:$project.version",
                 "$project.group:workflows:$project.version"
