+++ negotiation-cordapp/contracts/build.gradle
@@ -16,6 +16,7 @@ cordapp {
         vendor "Corda Open Source"
         targetPlatformVersion corda_platform_version.toInteger()
         minimumPlatformVersion corda_platform_version.toInteger()
+        versionId 1
     }
 }
