+++ negotiation-cordapp/workflows/build.gradle
@@ -17,6 +17,7 @@ cordapp {
         vendor "Corda Open Source"
         targetPlatformVersion corda_platform_version.toInteger()
         minimumPlatformVersion corda_platform_version.toInteger()
+        versionId 1
     }
 }
