+++ negotiation-cordapp/workflows/src/main/kotlin/negotiation/workflows/AcceptanceFlow.kt
@@ -1,39 +1,42 @@
-package com.negotiation
+package negotiation.workflows
 
 import co.paralleluniverse.fibers.Suspendable
+import negotiation.contracts.ProposalAndTradeContract
+import negotiation.contracts.ProposalState
+import negotiation.contracts.TradeState
 import net.corda.core.contracts.Command
 import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.flows.*
-import net.corda.core.identity.Party
 import net.corda.core.node.services.queryBy
 import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.utilities.ProgressTracker
 
-object ProposalFlow {
+object AcceptanceFlow {
     @InitiatingFlow
     @StartableByRPC
-    class Initiator(val isBuyer: Boolean, val amount: Int, val counterparty: Party) : FlowLogic<UniqueIdentifier>() {
+    class Initiator(val proposalId: UniqueIdentifier) : FlowLogic<Unit>() {
         override val progressTracker = ProgressTracker()
 
         @Suspendable
-        override fun call(): UniqueIdentifier {
+        override fun call() {
+            // Retrieving the input from the vault.
+            val inputCriteria = QueryCriteria.LinearStateQueryCriteria(linearId = listOf(proposalId))
+            val inputStateAndRef = serviceHub.vaultService.queryBy<ProposalState>(inputCriteria).states.single()
+            val input = inputStateAndRef.state.data
+
             // Creating the output.
-            val (buyer, seller) = when {
-                isBuyer -> ourIdentity to counterparty
-                else -> counterparty to ourIdentity
-            }
-            val output = ProposalState(amount, buyer, seller, ourIdentity, counterparty)
+            val output = TradeState(input.amount, input.buyer, input.seller, input.linearId)
 
             // Creating the command.
-            val commandType = ProposalAndTradeContract.Commands.Propose()
-            val requiredSigners = listOf(ourIdentity.owningKey, counterparty.owningKey)
-            val command = Command(commandType, requiredSigners)
+            val requiredSigners = listOf(input.proposer.owningKey, input.proposee.owningKey)
+            val command = Command(ProposalAndTradeContract.Commands.Accept(), requiredSigners)
 
             // Building the transaction.
-            val notary = serviceHub.networkMapCache.notaryIdentities.first()
+            val notary = inputStateAndRef.state.notary
             val txBuilder = TransactionBuilder(notary)
+            txBuilder.addInputState(inputStateAndRef)
             txBuilder.addOutputState(output, ProposalAndTradeContract.ID)
             txBuilder.addCommand(command)
 
@@ -41,12 +44,12 @@ object ProposalFlow {
             val partStx = serviceHub.signInitialTransaction(txBuilder)
 
             // Gathering the counterparty's signature.
+            val counterparty = if (ourIdentity == input.proposer) input.proposee else input.proposer
             val counterpartySession = initiateFlow(counterparty)
             val fullyStx = subFlow(CollectSignaturesFlow(partStx, listOf(counterpartySession)))
 
             // Finalising the transaction.
-            val finalisedTx = subFlow(FinalityFlow(fullyStx, listOf(counterpartySession)))
-            return finalisedTx.tx.outputsOfType<ProposalState>().single().linearId
+            subFlow(FinalityFlow(fullyStx, listOf(counterpartySession)))
         }
     }
 
@@ -56,7 +59,11 @@ object ProposalFlow {
         override fun call() {
             val signTransactionFlow = object : SignTransactionFlow(counterpartySession) {
                 override fun checkTransaction(stx: SignedTransaction) {
-                    // No checking to be done.
+                    val ledgerTx = stx.toLedgerTransaction(serviceHub, false)
+                    val proposee = ledgerTx.inputsOfType<ProposalState>().single().proposee
+                    if (proposee != counterpartySession.counterparty) {
+                        throw FlowException("Only the proposee can accept a proposal.")
+                    }
                 }
             }
