+++ negotiation-cordapp/workflows/src/main/kotlin/negotiation/workflows/ModificationFlow.kt
@@ -1,10 +1,11 @@
-package com.negotiation
+package negotiation.workflows
 
 import co.paralleluniverse.fibers.Suspendable
+import negotiation.contracts.ProposalAndTradeContract
+import negotiation.contracts.ProposalState
 import net.corda.core.contracts.Command
 import net.corda.core.contracts.UniqueIdentifier
 import net.corda.core.flows.*
-import net.corda.core.identity.Party
 import net.corda.core.node.services.queryBy
 import net.corda.core.node.services.vault.QueryCriteria
 import net.corda.core.transactions.SignedTransaction
