+++ negotiation-cordapp/workflows/src/test/kotlin/negotiation/workflows/contracts/AcceptanceContractTests.kt
@@ -1,8 +1,8 @@
-package com.negotiation.contracts
+package negotiation.workflows.contracts
 
-import com.negotiation.ProposalAndTradeContract
-import com.negotiation.ProposalState
-import com.negotiation.TradeState
+import negotiation.contracts.ProposalAndTradeContract
+import negotiation.contracts.ProposalState
+import negotiation.contracts.TradeState
 import net.corda.core.identity.CordaX500Name
 import net.corda.testing.core.DummyCommandData
 import net.corda.testing.core.TestIdentity
@@ -25,18 +25,18 @@ class AcceptanceContractTests {
 
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                 tweak {
-                    input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                    input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                    input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                    input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
                     fails()
                 }
                 tweak {
 
-                    output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
-                    output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                     fails()
                 }
-                input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 verifies()
             }
         }
@@ -48,15 +48,15 @@ class AcceptanceContractTests {
             transaction {
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                 tweak {
-                    input(ProposalAndTradeContract.ID,TradeState(1, alice.party, bob.party))
+                    input(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                     fails()
                 }
                 tweak {
-                    output(ProposalAndTradeContract.ID, ProposalState(1,alice.party,bob.party, alice.party, bob.party))
+                    output(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
                     fails()
                 }
-                input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 verifies()
             }
         }
@@ -67,14 +67,9 @@ class AcceptanceContractTests {
         ledgerServices.ledger {
             transaction {
                 input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 tweak {
                     command(listOf(alice.publicKey, bob.publicKey), DummyCommandData)
-                    fails()
-                }
-                tweak {
-                    command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
-                    command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                     fails()
                 }
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
@@ -90,16 +85,16 @@ class AcceptanceContractTests {
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                 tweak {
                     input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                    output(ProposalAndTradeContract.ID,TradeState(2,alice.party,bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(2, alice.party, bob.party))
                     fails()
                 }
                 tweak {
                     input(ProposalAndTradeContract.ID, ProposalState(2, alice.party, bob.party, alice.party, bob.party))
-                    output(ProposalAndTradeContract.ID,TradeState(1,alice.party,bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                     fails()
                 }
                 input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID,TradeState(1,alice.party,bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 verifies()
             }
         }
@@ -109,19 +104,19 @@ class AcceptanceContractTests {
     fun `buyer and seller are unmodified in the output`() {
         ledgerServices.ledger {
             transaction {
-                input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                 tweak {
-                    output(ProposalAndTradeContract.ID, TradeState(1,alice.party, charlie.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, alice.party, charlie.party))
                     fails()
                 }
                 tweak {
-                    output(ProposalAndTradeContract.ID, TradeState(1,charlie.party, bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, charlie.party, bob.party))
                     fails()
                 }
                 tweak {
-                    output(ProposalAndTradeContract.ID, TradeState(1,bob.party, bob.party))
+                    output(ProposalAndTradeContract.ID, TradeState(1, bob.party, bob.party))
                     fails()
                 }
                 verifies()
@@ -133,8 +128,8 @@ class AcceptanceContractTests {
     fun `proposal acceptance transactions have two required signers - the proposer and the proposee`() {
         ledgerServices.ledger {
             transaction {
-                input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 tweak {
                     command(listOf(alice.publicKey, charlie.publicKey), ProposalAndTradeContract.Commands.Accept())
                     fails()
@@ -153,8 +148,8 @@ class AcceptanceContractTests {
     fun `proposal acceptance transactions have no timestamp`() {
         ledgerServices.ledger {
             transaction {
-                input(ProposalAndTradeContract.ID,ProposalState(1, alice.party, bob.party, alice.party, bob.party))
-                output(ProposalAndTradeContract.ID, TradeState(1,alice.party,bob.party))
+                input(ProposalAndTradeContract.ID, ProposalState(1, alice.party, bob.party, alice.party, bob.party))
+                output(ProposalAndTradeContract.ID, TradeState(1, alice.party, bob.party))
                 command(listOf(alice.publicKey, bob.publicKey), ProposalAndTradeContract.Commands.Accept())
                 tweak {
                     timeWindow(Instant.now())
